# src/doc/swagger.yaml
openapi: 3.0.4
info:
  title: Blog API
  version: 2.0.0
  description: "API documentation for personal blogging platform"
servers:
  - url: http://localhost:5000/api
    description: Local development server
...
paths:
...
  /posts:
    get:
      operationId: getPosts
      summary: Get a list of posts
      description:  >
        Retrieve a paginated list of blog posts. You can control the number of posts returned using the `limit` parameter, 
        navigate through pages using the `cursor`, sort the results using the `sort` parameter, 
        search by keywords using `search`, and filter posts by tags using a comma-separated list in the `tags` parameter.
      tags: [Posts]
      parameters:
        - in: query
          name: limit
          schema:
            type: integer
            minimum: 1
            default: 9
          description: Max number of posts to return

        - $ref: "#/components/parameters/CursorParam"

        - $ref: "#/components/parameters/sortParam"

        - in: query
          name: search
          schema:
            type: string
            default: ""
          description: Search term to filter posts by title or description

        - in: query
          name: tags
          schema:
            type: string
          description: Comma-separated list of tags to filter posts
          example: "JavaScript,Node.js,API"

      responses:
        200:
          description: A list of posts
          content:
            application/json:
              schema:
                type: object
                properties:
                  cursor:
                    type: string
                    description: Cursor for pagination
                  posts:
                    type: array
                    items:
                      type: object
                      properties:
                        _id:
                          type: string
                          description: Unique identifier for the post
                        title:
                          type: string
                          description: Title of the post
                        slug:
                          type: string
                          description: Slugified title, used for URLs
                        cover:
                          type: string
                          description: URL of the cover image
                        tags:
                          type: array
                          items:
                            type: string
                          description: Tags associated with the post
                        likes:
                          type: integer
                          default: 0
                          description: Number of likes on the post
                        comments:
                          type: integer
                          default: 0
                          description: Number of comments on the post
                        createdAt:
                          type: string
                          format: date-time
                          description: Timestamp of post creation
                      
              example:
                cursor: 66e038559b9bb0ffc46836cf
                posts:
                  - _id: 64f0341f3df05e12a3c3a401
                    title: Understanding JavaScript Closures
                    slug: understanding-javascript-closures
                    cover: https://cdn.example.com/images/js-closures.png
                    tags: 
                      - JavaScript
                      - Closures
                      - Tutorial
                    likes: 0
                    comments: 0
                    createdAt: 2024-11-08T17:36:14.314Z
        400:
          description: Bad Request, Invalid query parameter
          content:
            application/json:
              schema:
                type: object
                properties:
                  limit:
                    type: string
                    description: Error message for invalid limit parameter
                    example: "Limit must be a positive integer"
                    nullable: true
                  cursor:
                    type: string
                    description: Error message for invalid cursor parameter
                    example: "Invalid cursor format"
                    nullable: true
                  sort:
                    type: string
                    description: Error message for invalid sort parameter
                    example: "Invalid sort value, must be one of: latest, oldest, top"
                    nullable: true
                  search:
                    type: string
                    description: Error message for invalid search parameter 
                    example: "Search term must be a string"
                    nullable: true
                  tags:
                    type: string
                    description: Error message for invalid tags parameter
                    example: "Tags must be a comma-separated list of strings"
                    nullable: true
        500:
          $ref: "#/components/responses/ServerError"

    post:
      operationId: createPost
      summary: Create a new post
      description: >
        Allows an authenticated user with admin or moderator privileges to create a new blog post. 
        Requires title, description, headers, cover image URL, content, and a list of tags.
      tags: [Posts]
      security:
        - JWTAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - title
                - description
                - headers
                - cover
                - content
                - tags
              properties:
                title:
                  type: string
                  description: Title of the post
                description:
                  type: string
                  description: Short summary or subtitle of the post
                headers:
                  type: array
                  items:
                    $ref: "#/components/schemas/PostHeader"
                cover:
                  type: string
                  description: URL of the cover image
                content:
                  type: array
                  items:
                    $ref: "#/components/schemas/PostContent"
                tags:
                  type: array
                  description: List of tags associated with the post
                  items:
                    type: string
            example:
              title: "Node.js Streams Explained"
              description: "An introduction to readable and writable streams"
              headers:
                  - id: 5a52da70-507c-497d-9bd7-968be323c912
                    type: H2
                    value: Introduction
                  - id: 5a52da70-507c-497d-9bd7-968be32dc912
                    type: H2
                    value: What are Streams?
              cover: "https://cdn.example.com/images/node-streams.png"
              content:
                - type: Image
                  value: https://cdn.example.com/images/js-closures.png
                - type: Editor
                  value: "Closures are a fundamental concept..."
                - type: Code Snippet
                  value: "function outer() { ... }"
                  language: "javascript"
              tags: [JavaScript, Closures, Tutorial]
      responses:
        201:
          $ref: "#/components/responses/Created"
        400:
          description: Bad Request – Invalid input fields
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: object
                    properties:
                      title:
                        type: string
                        nullable: true
                      description:
                        type: string
                        nullable: true
                      headers:
                        type: string
                        nullable: true
                      cover:
                        type: string
                        nullable: true
                      content:
                        type: string
                        nullable: true
                      tags:
                        type: string
                        nullable: true
                example:
                  message:
                    title: "Title is required"
                    description: null
                    headers: null
                    cover: null
                    content: null
                    tags: "Tags must be an array of strings"
        401:
          $ref: "#/components/responses/Unauthorized"
        403:
          $ref: "#/components/responses/Forbidden"
        500:
          $ref: "#/components/responses/ServerError"
...
  /posts/{postId}/likes:
    post:
      operationId: likePost
      summary: Like a post
      description: >
        Allows an authenticated user to like a blog post. If the post has already been liked by the same user, 
        the API will respond with a message indicating the post was already liked.
      tags: [Posts]
      security:
        - JWTAuth: []
      parameters:
        - $ref: "#/components/parameters/PostIdParam"
      responses:
        200:
          $ref: "#/components/responses/Success"
        400:
          description: Bad Request – Invalid post ID
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: Invalid post id
        401:
          $ref: "#/components/responses/Unauthorized"
        404:
          $ref: "#/components/responses/PostNotFound"
        500:
          $ref: "#/components/responses/ServerError"

    delete:
      operationId: unlikePost
      summary: Remove a like from a post
      description: >
        Removes a like from a post by the authenticated user.  
        If the user hasn't liked the post, a 200 response is returned with a message.
      tags: [Posts]
      security:
        - JWTAuth: []
      parameters:
        - $ref: "#/components/parameters/PostIdParam"
      responses:
        200:
          $ref: "#/components/responses/Success"
        400:
          description: Invalid post ID
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "Invalid post id"
        401:
          $ref: "#/components/responses/Unauthorized"
        404:
          $ref: "#/components/responses/PostNotFound"
        500:
          $ref: "#/components/responses/ServerError"
  /posts/{postId}:
    patch:
      operationId: updatePost
      summary: Update a post
      description: >
        Update an existing post. Only users with the `ADMIN` or `MODERATOR` role can access this route.
        If no valid fields are provided, it will return 204 No Content. If a token is refreshed, it returns a new access token.
      tags: [Posts]
      security:
        - JWTAuth: []
      parameters:
        - $ref: "#/components/parameters/PostIdParam"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                title:
                  type: string
                description:
                  type: string
                headers:
                  type: array
                  items:
                    $ref: "#/components/schemas/PostHeader"
                cover:
                  type: string
                  format: uri
                content:
                  type: array
                  items:
                    $ref: "#/components/schemas/PostContent"
                tags:
                  type: array
                  items:
                    type: string
              example:
                title: "Updated Post Title"
                description: "Updated description of the post"
                headers:
                  - id: "h1"
                    type: "H1"
                    value: "Intro"
                cover: "https://cdn.example.com/newcover.png"
                content:
                  - type: "Editor"
                    value: "Updated content here..."
                tags: ["webdev", "typescript"]
      responses:
        200:
          $ref: "#/components/responses/Updated"
        400:
          description: Invalid request body or post ID
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: object
                    additionalProperties:
                      type: string
                      nullable: true
                example:
                  message:
                    title: "Title must be at least 5 characters"
                    description: null
                    cover: "Cover must be a valid URL"
        401:
          $ref: "#/components/responses/Unauthorized"
        403:
          $ref: "#/components/responses/Forbidden"
        404:
          $ref: "#/components/responses/PostNotFound"
        500:
          $ref: "#/components/responses/ServerError"

    delete:
      operationId: deletePost
      summary: Delete a post
      description: >
        Deletes a post and all its associated comments and likes.  
        Only `ADMIN` or `MODERATOR` roles can perform this operation.
      tags: [Posts]
      security:
        - JWTAuth: []
      parameters:
        - $ref: "#/components/parameters/PostIdParam"
      responses:
        200:
          $ref: "#/components/responses/Deleted"
        400:
          description: Invalid post ID
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "Invalid post id"
        401:
          $ref: "#/components/responses/Unauthorized"
        403:
          $ref: "#/components/responses/Forbidden"
        404:
          $ref: "#/components/responses/PostNotFound"
        500:
          $ref: "#/components/responses/ServerError"
...
components:
  securitySchemes:
    JWTAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: >
        Use the JWT token obtained from login or registration to access protected routes.
  parameters:
    PostIdParam:
      in: path
      name: postId
      required: true
      schema:
        type: string
        format: objectId
        example: "64f0341f3df05e12a3c3a401"
      description: ID of the post
...
    CursorParam:
      in: query
      name: cursor
      schema:
        type: string
        nullable: true
        example: 64f0341f3df05e12a3c3a401
      description: Cursor to the next page (null if no more)
    sortParam:
      in: query
      name: sort
      schema:
        type: string
        enum: [latest, oldest, top]
        default: latest
      description: Sort order
  responses:
...
    Success:
      description: Success, if the user's access token has expired a new one is generated.
      content:
        application/json:
          schema:
            type: object
            properties:
              message:
                type: string
                description: Response message
                example: Success
              accessToken:
                type: string
                nullable: true
                description: JWT access token for the user
                example: "eyJhbGciOiJIUzI1NiIsInR..."
    Created:
      description: Created, if the user's access token has expired a new one is generated.
      content:
        application/json:
          schema:
            type: object
            properties:
              message:
                type: string
                description: Response message
                example: Created
              accessToken:
                type: string
                nullable: true
                description: JWT access token for the user
                example: "eyJhbGciOiJIUzI1NiIsInR..."
    Updated:
      description: Updated, if the user's access token has expired a new one is generated.
      content:
        application/json:
          schema:
            type: object
            properties:
              message:
                type: string
                description: Response message
                example: Updated
              accessToken:
                type: string
                nullable: true
                description: JWT access token for the user
                example: "eyJhbGciOiJIUzI1NiIsInR..."
    Deleted:
      description: Deleted, if the user's access token has expired a new one is generated.
      content:
        application/json:
          schema:
            type: object
            properties:
              message:
                type: string
                description: Response message
                example: Deleted
              accessToken:
                type: string
                nullable: true
                description: JWT access token for the user
                example: "eyJhbGciOiJIUzI1NiIsInR..."
...
    PostNotFound:
      description: Post not found
      content:
        application/json:
          schema:
            type: object
            properties:
              message:
                type: string
                description: Error message for post not found
                example: "Post not found"
...
    Unauthorized:
      description: Unauthorized – user not authenticated or token is invalid
      content:
        application/json:
          schema:
            type: object
            properties:
              message:
                type: string
                example: Unauthorized
    Forbidden:
      description: Forbidden – user lacks permission (read description to know what roles are authorized)
      content:
        application/json:
          schema:
            type: object
            properties:
              message:
                type: string
                example: Forbidden
    ServerError:
      description: Internal server error
      content:
        application/json:
          schema:
            type: object
            properties:
              message:
                type: string
                description: Error message
  schemas:
    PostHeader:
      type: object
      properties:
        id:
          type: string
          description: Unique identifier for the header block
        type:
          type: string
          enum: [H2, H3]
          description: The type of header (H2 or H3)
        value:
          type: string
          description: The text of the header
    PostContent:
      type: object
      properties:
        type:
          type: string
          enum: [Image, Code Snippet, Code Output, Editor]
          description: The block type
        value:
          type: string
          description: The actual content or reference
        language:
          type: string
          description: Language used (only for code snippets)
          nullable: true